requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.1
fsspec==2025.5.1
gradio==5.33.0
gradio-client==1.10.2
groovy==0.1.2
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
jinja2==3.1.6
kiwisolver==1.4.8
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnx==1.18.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.5
pydantic-core==2.33.2
pydub==0.25.1
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
ruff==0.11.13
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.15.3
seaborn==0.13.2
segmentation-models-pytorch==0.5.0
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
timm==1.0.15
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.151
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
werkzeug==3.1.3
